Fix Open/Closed Principal violation in CommandLineUI
CommandLineUI had details about how a particular type of player
should be managed.  I've changed the code such that the Player
is asked for it's move which then either uses an AI algorithm in
the ComputerPlayer's case or asks the UI for a move.

This felt a bit like going back in time as I had removed the need
for the Player to interract with the UI as it wasn't necessary.

I had a comment in my code tests that reminded me that I was in
two minds as to which was the better option: 1) have the UI be responsible
for prompting for moves or 2) have the Player ask the UI for a
move if it is necessary.   It was a trade off, but I hadn't thought
about the OCP violation at the time and so went for the former.  I
now think it is better this way.  Additionally, this way I think
it will be better for the GUI UI.